Update and simplify gcode handling

+++ src/geometry.h
@@ -29,7 +29,7 @@ void moveArcballTrans(Matrix4fT &matfT, const Vector3d &delta);
 void setArcballTrans(Matrix4fT &matfT, const Vector3d &trans);
 void rotArcballTrans(Matrix4fT &transform,  const Vector3d &axis, double angle);
 
-long double planeAngleBetween(const Vector2d &V1, const Vector2d &V2);
+double planeAngleBetween(const Vector2d &V1, const Vector2d &V2);
 
 double dist3D_Segment_to_Segment(const Vector3d &S1P0, const Vector3d &S1P1,
 				 const Vector3d &S2P0, const Vector3d &S2P1, double SMALL_NUM);
